+++ src/Mayaqua/HTTP.h
@@ -1,0 +1,13 @@
+#ifndef HTTP_H
+#define HTTP_H
+
+// MIME type
+struct HTTP_MIME_TYPE
+{
+	char *Extension;
+	char *MimeType;
+};
+
+char *GetMimeTypeFromFileName(char *filename);
+
+#endif
